all_bot/bot_type/requirements.txt
python-dotenv
langchain-community
openai
python-socketio 